simplify config changes further
